+++ src/agentlang/cli/command.clj
@@ -1,13 +1,13 @@
-(ns fractl.cli.command
+(ns agentlang.cli.command
   (:require [cemerick.pomegranate.aether :as aether]
             [clojure.edn :as edn]
             [clojure.java.io :as io]
             [clojure.string :as string]
             [clojure.pprint :as pp]
             [clojure.walk :as walk]
-            [fractl.cli.core :as core]
-            [fractl.cli.newproj :as newproj]
-            [fractl.cli.util :as util])
+            [agentlang.cli.core :as core]
+            [agentlang.cli.newproj :as newproj]
+            [agentlang.cli.util :as util])
   (:import (java.io StringWriter)))
 
 
@@ -56,27 +56,27 @@
     "app"      (if-let [error (util/invalid-project-name? new-name)]
                  (util/err-exit "Invalid app name:" error)
                  (do
-                   (util/err-println "Creating new Fractl app")
+                   (util/err-println "Creating new AgentLang app")
                    (newproj/create-new-app new-name)))
     "resolver" (if-let [error (util/invalid-project-name? new-name)]
                  (util/err-exit "Invalid resolver name:" error)
                  (do
-                   (util/err-println "Creating new Fractl resolver")
+                   (util/err-println "Creating new Agentlang resolver")
                    (util/err-exit "Not yet implemented")))))
 
 
-(defn command-fractl [dirname msg-prefix fractl-command args]
+(defn command-agentlang [dirname msg-prefix agentlang-command args]
   (let [app-model      (core/read-model dirname)
         app-version    (:version app-model "(unknown app version)")
-        fractl-version (core/rewrite-fractl-version (:fractl-version app-model))
+        agentlang-version (core/rewrite-agentlang-version (:agentlang-version app-model))
         classpath (-> app-model
                       core/find-dependencies
                       core/fetch-dependencies
                       core/prepare-classpath)]
-    (util/err-println (format "%s %s with Fractl %s"
+    (util/err-println (format "%s %s with AgentLang %s"
                               msg-prefix
-                              app-version fractl-version))
-    (core/run-fractl dirname classpath fractl-command args)))
+                              app-version agentlang-version))
+    (core/run-agentlang dirname classpath agentlang-command args)))
 
 
 (defn command-clone [[command repo-uri & args]]
@@ -99,7 +99,7 @@
                           core/find-dependencies
                           core/fetch-dependencies
                           core/prepare-classpath)]
-        (core/run-fractl repo-name classpath command args))
+        (core/run-agentlang repo-name classpath command args))
       git-result)))
 
 
@@ -108,12 +108,12 @@
                         slurp
                         (edn/read-string)
                         :version)
-        fractl-version (when-let [model (binding [*err* (StringWriter.)]
-                                          (core/read-model core/current-directory))]
-                         (:fractl-version model))
+        agentlang-version (when-let [model (binding [*err* (StringWriter.)]
+                                             (core/read-model core/current-directory))]
+                            (:agentlang-version model))
         clijvm-version (System/getProperty "java.version")
         version {:cli-version cliapp-version
-                 :fractl-version fractl-version
+                 :agentlang-version agentlang-version
                  :jvm-version clijvm-version}]
     (case (-> version-format
               str
@@ -121,33 +121,33 @@
               string/trim)
       "edn" (pp/pprint version)
       "json" (printf "{\"cli-version\": \"%s\",
- \"fractl-version\": %s,
+ \"agentlang-version\": %s,
  \"jvm-version\": \"%s\"}\n"
                      cliapp-version
-                     (and fractl-version (format "\"%s\"" fractl-version))
+                     (and agentlang-version (format "\"%s\"" agentlang-version))
                      clijvm-version)
       (do
         (println "CLI version:" cliapp-version)
-        (println "Fractl version:" (or fractl-version "Unavailable"))
+        (println "AgentLang version:" (or agentlang-version "Unavailable"))
         (println "JVM version:" clijvm-version)))
     (flush)))
 
 
 (defn command-help []
   (binding [*out* *err*]
-    (util/err-println "Syntax: ftl <command> [command-args]
+    (util/err-println "Syntax: agent <command> [command-args]
 
-ftl deps               Fetch dependencies for a Fractl app
-ftl depstree           Print dependency-tree for a Fractl app
-ftl classpath          Print classpath for a Fractl app
-ftl clonenrepl         Clone a (Git) repo and start nREPL server in the app
-ftl clonerepl          Clone a (Git) repo and start REPL in the app
-ftl clonerun           Clone a (Git) repo and run the app
-ftl new app <ap-name>  Create a new Fractl app
-ftl nrepl              Start an nREPL server
-ftl repl               Start a local REPL
-ftl run [run-args]     Run a Fractl app
-ftl version [format]   Print ftl version (format: edn/json)")))
+agent deps               Fetch dependencies for an AgentLang app
+agent depstree           Print dependency-tree for an AgentLang app
+agent classpath          Print classpath for an AgentLang app
+agent clonenrepl         Clone a (Git) repo and start nREPL server in the app
+agent clonerepl          Clone a (Git) repo and start REPL in the app
+agent clonerun           Clone a (Git) repo and run the app
+agent new app <ap-name>  Create a new AgentLang app
+agent nrepl              Start an nREPL server
+agent repl               Start a local REPL
+agent run [run-args]     Run an AgentLang app
+agent version [format]   Print agentlang.cli version (format: edn/json)")))
 
 
 (defn process-command
@@ -161,15 +161,15 @@ ftl version [format]   Print ftl version (format: edn/json)")))
                  "clonerun" (command-clone (cons "run" args))
                  "help" (command-help)
                  "new" (command-new args)
-                 "nrepl" (command-fractl core/current-directory
-                                         "Starting nREPL server for app"
-                                         "nrepl" args)
-                 "repl" (command-fractl core/current-directory
-                                        "Starting REPL for app"
-                                        "repl" args)
-                 "run" (command-fractl core/current-directory
-                                       "Starting app"
-                                       "run" args)
+                 "nrepl" (command-agentlang core/current-directory
+                                            "Starting nREPL server for app"
+                                            "nrepl" args)
+                 "repl" (command-agentlang core/current-directory
+                                           "Starting REPL for app"
+                                           "repl" args)
+                 "run" (command-agentlang core/current-directory
+                                          "Starting app"
+                                          "run" args)
                  "version" (command-version args)
                  (do
                    (if (nil? command)