+++ test/agentlang/cli/command_test.clj
@@ -1,13 +1,13 @@
-(ns fractl.cli.command-test
+(ns agentlang.cli.command-test
   (:require [clojure.test :refer [deftest is testing]]
-            [fractl.cli.command :as command]
-            [fractl.cli.core :as core])
+            [agentlang.cli.command :as command]
+            [agentlang.cli.core :as core])
   (:import (clojure.lang ExceptionInfo)))
 
 
 (deftest test-command-deps
-  (testing "Missing model.fractl file"
+  (testing "Missing model.al file"
     (is (= 1 (command/command-deps "."))))
-  (testing "Unspecified Fractl version in model.fractl"
+  (testing "Unspecified AgentLang version in model.al"
     (with-redefs [core/read-model (fn [_] {})]
       (is (seq (command/command-deps "."))))))