+++ CHANGELOG.md
@@ -6,6 +6,15 @@ All notable changes to this project will be documented in this file. This change
 2. Support for multiple source paths (`:model-paths` in `model.fractl`)
 3. Resolver template, using `ftl new resolver <resolver-name>`
 
+## [0.3.0] - 2024-09-??
+
+- Rename project to agentlang.cli
+  - Rename script to `agent`
+  - Refactor/rename internal source/test files
+- Target AgentLang (rather than Fractl)
+  - Baseline version `0.6.0-alpha2`
+- Update Clojure dependency to version 1.12.0
+
 ## [0.2.0] - 2024-09-05
 
 - Cloning code repo and running as one command